Add image below second button
Inserted an image component below the second button in the relevant page.
[skip gpt_engineer]
