+++ src/pages/pt/steps/WelcomeStepPT.tsx
@@ -1,102 +1,17 @@
 import { motion } from "framer-motion";
 import { Button } from "@/components/ui/button";
-import { useEffect, useRef } from "react";
-import { BenefitBlock } from "@/components/marketing/landing/BenefitBlock";
+import { WelcomeHeader } from "@/components/marketing/landing/welcome/WelcomeHeader";
+import { ProductScroll } from "@/components/marketing/landing/welcome/ProductScroll";
+import { BenefitsScroll } from "@/components/marketing/landing/welcome/BenefitsScroll";
 
 interface WelcomeStepProps {
   onNext: () => void;
 }
 
 export const WelcomeStepPT = ({ onNext }: WelcomeStepProps) => {
-  const scrollRef = useRef<HTMLDivElement>(null);
-  const imagesScrollRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const scrollContainer = scrollRef.current;
-    const imagesContainer = imagesScrollRef.current;
-    
-    if (!scrollContainer || !imagesContainer) return;
-
-    const scroll = () => {
-      if (scrollContainer.scrollLeft >= scrollContainer.scrollWidth / 2) {
-        scrollContainer.scrollLeft = 0;
-      } else {
-        scrollContainer.scrollLeft += 0.5;
-      }
-    };
-
-    const scrollImages = () => {
-      if (imagesContainer.scrollLeft >= imagesContainer.scrollWidth / 2) {
-        imagesContainer.scrollLeft = 0;
-      } else {
-        imagesContainer.scrollLeft += 0.5;
-      }
-    };
-
-    const intervalId = setInterval(scroll, 16);
-    const imagesIntervalId = setInterval(scrollImages, 16);
-
-    return () => {
-      clearInterval(intervalId);
-      clearInterval(imagesIntervalId);
-    };
-  }, []);
-
-  const productImages = [
-    "/lovable-uploads/e64b31f9-2f7c-41d0-ae9d-f805f40571d7.png",
-    "/lovable-uploads/b5a0ae27-4415-49cc-a3f0-05b550b23869.png",
-    "/lovable-uploads/9954d0ec-cdbf-439d-8dc7-52e39fc08778.png",
-    "/lovable-uploads/e82547dd-1abc-486c-8932-56bacd4b77bc.png",
-    "/lovable-uploads/02151e6b-cad9-45a1-8115-97fd85ff7aae.png",
-    "/lovable-uploads/9ce98c58-429f-40c2-b6fd-7dc44224e3dc.png",
-    "/lovable-uploads/516b2ea6-8b74-4c84-b3fd-a91636e32f9b.png",
-  ];
-
-  const benefitBlocks = [
-    {
-      title: "Sob Demanda",
-      description: "Comece sua marca sem o peso dos custos de inventário"
-    },
-    {
-      title: "Fornecedores Americanos",
-      description: "Trabalhe com fabricantes confiáveis e focados em qualidade"
-    },
-    {
-      title: "+250 Produtos Disponíveis",
-      description: "Amplo catálogo de produtos para você escolher e personalizar sua marca"
-    },
-    {
-      title: "Suporte Dedicado",
-      description: "Equipe especializada para ajudar em cada etapa"
-    },
-    {
-      title: "Entrega Rápida",
-      description: "Logística eficiente para seus produtos"
-    },
-    {
-      title: "Qualidade Premium",
-      description: "Produtos de alta qualidade para sua marca"
-    }
-  ];
-
   return (
     <div className="text-center space-y-6 w-full max-w-lg mx-auto px-3 pt-8 sm:pt-8">
-      <motion.div
-        initial={{ opacity: 0, y: 20 }}
-        animate={{ opacity: 1, y: 0 }}
-        className="space-y-3"
-      >
-        <div className="flex flex-col items-center gap-1 mt-8 sm:mt-6">
-          <img
-            src="https://assets.cdn.filesafe.space/Q5OD6tvJPFLSMWrJ9Ent/media/673c037af980e11b5682313e.png"
-            alt="GetBrands Logo"
-            className="h-8 sm:h-10 mb-2"
-          />
-          <h2 className="text-3xl sm:text-4xl md:text-5xl font-bold text-gray-900 leading-tight">
-            Crie Sua Marca Própria nos EUA Sem Investir em Estoque
-          </h2>
-        </div>
-      </motion.div>
+      <WelcomeHeader />
 
       <motion.div
         initial={{ opacity: 0, y: 20 }}
@@ -104,26 +19,7 @@ export const WelcomeStepPT = ({ onNext }: WelcomeStepProps) => {
         transition={{ delay: 0.2 }}
         className="w-full -mx-3 sm:mx-0 mt-8"
       >
-        <div 
-          ref={imagesScrollRef}
-          className="overflow-hidden relative w-screen sm:w-auto"
-          style={{ WebkitOverflowScrolling: 'touch' }}
-        >
-          <div className="flex gap-4 sm:gap-6 animate-scroll w-[200%] justify-center">
-            {[...productImages, ...productImages].map((image, index) => (
-              <div 
-                key={`${index}`}
-                className="flex-none w-[220px] sm:w-[280px] aspect-square bg-white rounded-lg p-4"
-              >
-                <img 
-                  src={image} 
-                  alt={`Product ${index + 1}`}
-                  className="w-full h-full object-contain"
-                />
-              </div>
-            ))}
-          </div>
-        </div>
+        <ProductScroll />
       </motion.div>
 
       <motion.div
@@ -172,33 +68,7 @@ export const WelcomeStepPT = ({ onNext }: WelcomeStepProps) => {
           transition={{ delay: 0.4 }}
           className="w-full -mx-3 sm:mx-0"
         >
-          <div 
-            ref={scrollRef}
-            className="overflow-hidden relative w-screen sm:w-auto"
-            style={{ WebkitOverflowScrolling: 'touch' }}
-          >
-            <div className="flex gap-2 sm:gap-3 animate-scroll w-[200%] justify-center">
-              {benefitBlocks.map((block, index) => (
-                <BenefitBlock
-                  key={`first-${index}`}
-                  title={block.title}
-                  description={block.description}
-                  keyPrefix="first"
-                  index={index}
-                />
-              ))}
-              
-              {benefitBlocks.map((block, index) => (
-                <BenefitBlock
-                  key={`second-${index}`}
-                  title={block.title}
-                  description={block.description}
-                  keyPrefix="second"
-                  index={index}
-                />
-              ))}
-            </div>
-          </div>
+          <BenefitsScroll />
         </motion.div>
 
         <p className="text-base sm:text-lg md:text-xl text-gray-600 max-w-md mx-auto">
@@ -211,6 +81,13 @@ export const WelcomeStepPT = ({ onNext }: WelcomeStepProps) => {
           >
             Ver Catálogo de Produtos
           </Button>
+          <div className="mt-4">
+            <img 
+              src="/lovable-uploads/08ee0a1f-fed9-4b8a-8381-acda14355625.png"
+              alt="FDA Approved"
+              className="h-16 mx-auto"
+            />
+          </div>
         </div>
         <div className="h-10"></div>
       </motion.div>
